Separate `CodeGen` from `CodeGenRunner`
