+++ language-support/java/codegen/BUILD.bazel
@@ -92,7 +92,6 @@ da_scala_test(
         "//daml-lf/interface",
         "//language-support/codegen-common",
         "//language-support/java/bindings:bindings-java",
-        "@maven//:ch_qos_logback_logback_classic",
         "@maven//:com_squareup_javapoet",
         "@maven//:org_scalatest_scalatest_compatible",
         "@maven//:org_slf4j_slf4j_api",
